dwc3: msm: Notify max power info if charging is enabled only
Currently driver is notifying maxpower info in case of host mode
for optimizing boost current by pmic. For secondary instance, where
charging is disabled, it is not required to notify max power. Else
this leads to functionality not working on primary USB ports where
charging is enabled. Hence fix it by notifying it to charger driver
only if charging is not disabled.

Change-Id: Ie07854d6169c10c6ef02f0ff503c9dea95844d6f
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>